requirements.txt
streamlit==1.13.0
tensorflow==2.10.0
numpy==1.21.4
pillow==8.4.0
matplotlib==3.5.1
